requirements.txt
black==23.3.0
mlflow==2.3
numpy==1.24.3
pyarrow==11.0.0
PyYAML==6.0
pandas==2.0.1
scikit-learn==1.2.2
xgboost==1.7.5

fastapi==0.95.1
pydantic==1.10.7
uvicorn==0.22.0

python-dotenv==1.0.0
pymongo==4.3.3
google-cloud-storage==2.9.0

